refactoring notifications routes and files
